Use SOCK_STREAM instead of SOCK_DGRAM for MessageChannel on macOS
OS X doesn't support SOCK_SEQPACKET, which is our default for MessageChannels on unix. Previously we used SOCK_DGRAM, but SOCK_DGRAM has the disadvantage that a compromised process can use it to send datagrams to other, unconnected sockets on the system. However it turns out OS X presevers message boundaries for unix sockets when sendmsg/recvmsg are used, so in this case we just use a standard SOCK_STREAM socketpair